9.1.4 allow to delete Link only for author

+++ app/controllers/links_controller.rb
@@ -1,6 +1,8 @@
 class LinksController < ApplicationController
+  before_action :authenticate_user!
+  
   def destroy
-    link.destroy
+    link.destroy if current_user.author_of?(link.linkable)
   end
 
   private
